Moved DATA to main_data.py
