+++ pxbuild/controll/from_pxmetadata_file.py
@@ -15,9 +15,11 @@ from pxbuild.models.output.pxfile.px_file_model import PXFileModel
 from pxbuild.models.output.agg_vs.vs_file_model import _VSFileModel
 from pxbuild.models.output.agg.agg_file_model import AggFileModel
 
+from .helpers.small_static_functions import Commons
+
 from .helpers.datadata_helpers.datadatasource import Datadatasource
 from .helpers.datadata_helpers.for_get_data import ForGetData
-from .helpers.datadata_helpers.data_formatter import DataFormatter
+from .helpers.datadata_helpers.main_data import MapData
 from .helpers.loaded_jsons import LoadedJsons 
 
 
@@ -85,10 +87,8 @@ class LoadFromPxmetadata:
             self.add_metaid_to_pxfile(out_model)
 
 
-            #fixdata = FixData()
-            #fixdata.get_data(out_model)
-            self.get_data(out_model)
-
+            fixdata = MapData(self._datadata, self._pxmetadata_model, self._config, self._for_get_data_by_varid, self._stub , self._heading)
+            fixdata.map_data(out_model)
 
 
 
@@ -106,96 +106,8 @@ class LoadFromPxmetadata:
 
         self.make_vs_file()
 
-    def calculate_factor(self) -> int:
-        variables_in_output_order = self.get_variable_list()
-        curr_factor = 1
-        for vari in reversed(variables_in_output_order):
-            temp_for_get_data: ForGetData = self._for_get_data_by_varid[vari]
-            temp_for_get_data.factor = curr_factor
-            prev_number = temp_for_get_data._length_of_codelist
-            curr_factor = curr_factor * prev_number
 
-        array_size = curr_factor
 
-        return array_size
-
-    def get_data(self, out_model: PXFileModel) -> None:
-        # /// MINDEX:
-        # /// We need to convert a point(one value for each variable) in
-        # /// the cube to a number(the index of the array).
-        # ///
-        # /// k,j, i... 1-based counters
-        # /// Nx number of values for x
-        # /// Factor_k=Nj*Ni
-        # /// Factor_j= Ni
-        # /// Factor_i = 1
-        # /// index = Factor_k*(k-1) + Factor_j*(j-1) + Factor_i(i-1) </remarks>
-        #  in python things are zero-based but the idea is the same
-        # /// </summary>
-        if not self._add_language_independent:
-           return
-        
-        start_get_data = time.time()
-
-        matrix_size = self.calculate_factor()
-
-        missing_row_symbol = self._pxmetadata_model.dataset.row_missing
-        missing_cell_symbol = self._pxmetadata_model.dataset.cell_missing
-
-        column_code_map = self.get_measurement_column_code_mapping()
-
-        start_tidy = time.time()
-        df = self._datadata.GetTidyDF(self._config.contvariable_code, column_code_map)
-
-        end_tidy = time.time()
-        time_used_tidy = end_tidy - start_tidy
-        print("Time: GetTidyDF:", time_used_tidy)
-
-        self.add_out_index(df)
-        self.add_out_value(missing_cell_symbol, df)
-        merged_df = self.add_missing_rows(matrix_size, missing_row_symbol, df)
-
-        out_data = merged_df["out_value"].tolist()
-
-        #merged_df = merged_df.sort_values(by=["out_index"])
-        formatter = DataFormatter(self._heading, self._for_get_data_by_varid)
-        number_of_columns_per_line = formatter.calculate_line_break()
-
-        out_model.data.set(out_data, number_of_columns_per_line)
-
-        end_get_data = time.time()
-        time_used_get_data = end_get_data - start_get_data
-        print("Time: GetData:", time_used_get_data)
-
-    def add_missing_rows(self, matrix_size, missing_row_symbol, df):
-        matrix_df = pd.DataFrame({"out_index": range(matrix_size)})
-
-        # Merge the two DataFrames
-        merged_df = pd.merge(matrix_df, df, on="out_index", how="left")
-
-        # Fill missing values with "MISSING"
-        merged_df["out_value"].fillna(missing_row_symbol, inplace=True)
-        return merged_df
-
-    def add_out_value(self, missing_cell_symbol, df):
-        conditions = [df["VALUE"] != "", df["SYMBOL"] != ""]
-
-        choices = [df["VALUE"].astype(str), df["SYMBOL"].astype(str)]
-
-        start = time.time()
-        df["out_value"] = np.select(conditions, choices, missing_cell_symbol)
-        end = time.time()
-        time_used = end - start
-        print("Time: numpy select in:", time_used)
-
-    def add_out_index(self, df):
-        columns_to_sum = []
-        for col in self._for_get_data_by_varid.values():
-            # Mapping the values using the dictionary
-            contrib_col_name = "int_" + col._colname_in_dataframe
-            df[contrib_col_name] = df[col._colname_in_dataframe].map(col._position_of_value) * col.factor
-            columns_to_sum.append(contrib_col_name)
-        df["out_index"] = df[columns_to_sum].sum(axis=1)
 
     def add_metaid_to_pxfile(self, out_model: PXFileModel) -> None:
         if self._add_language_independent:
@@ -205,16 +117,7 @@ class LoadFromPxmetadata:
         for vari in self._metaid_valiable:
             out_model.meta_id.set(" ".join(self._metaid_valiable[vari]), vari, None, self._current_lang)
 
-    def get_variable_list(self) -> List[str]:
-        return self._stub + self._heading
-
-    def get_measurement_column_code_mapping(self) -> dict:
-        column_code_map = {}
-        for measurement_var in self._pxmetadata_model.dataset.measurements:
-            column_code_map[measurement_var.column_name] = measurement_var.code
-
-        return column_code_map
-    
+   
 
     def map_aggregallowed_to_pxfile(self, out_model: PXFileModel):
         # Check if all values in the array are True
@@ -226,8 +129,7 @@ class LoadFromPxmetadata:
         lang = self._current_lang
         model = self._pxmetadata_model.dataset
 
-        vari_list = self.get_variable_list()
-
+        vari_list =  Commons.get_variable_list(self._stub, self._heading ) 
         tmp_string = ", ".join(vari_list[:-1])
 
         title = (
@@ -463,7 +365,7 @@ class LoadFromPxmetadata:
         out_model.source.set(in_config.source[self._current_lang], self._current_lang)
 
 
-
+    #################################
     # def makeVsFile(self,out_vs_model:_VSFileModel):
     def make_vs_file(self):
         if not self._pxmetadata_model.dataset.coded_dimensions:
